Trim leading slashes from watch paths (#534)
When globbing files for execution, we trim leading slashes from globs, so that everything is always relative to the package, instead of the root filesystem, even if you write e.g. `/src/**`. However, I forgot to do that normalization for those globs when we pass them to chokidar, so they wouldn't get watched.

Fixes https://github.com/google/wireit/issues/529