Prevent fingerprint size explosion by using hashes of dependency fingerprints (#546)
After each script runs, we write its JSON-stringified *fingerprint* to a file, so that the next time it runs we can compare the fingerprint to decide whether it needs to run again or not.

Part of a fingerprint is the fingerprint of its dependencies (this is how what we now describe as the "cascade" behavior works). (The full list of things in the fingerprint is at https://github.com/google/wireit#fingerprint).

Previously, when nesting dependency fingerprints, we nested the *entire data object for that dependency*. The problem with this is that it can cause an explosion of data in the fingerprint, which at a certain point will be too large for V8's `JSON.stringify` function to handle, causing either an `Invalid string length` error or a `heap out of memory` error. This also would have hurt performance, since we were serializing and writing unnecessarily large amounts of data.

This PR fixes that problem. Now, when embedding the fingerprints of dependencies, we instead embed a *hash* of the fingerprint, instead of the whole object. This bounds the size of any given fingerprint dependency to 64 bytes, whereas previously it could grow to any size -- in some configurations expontentially.

Fixes https://github.com/google/wireit/issues/543