Run all glob tests against chokidar too (#551)
This PR runs every glob test against chokidar, since we should have identical behavior in theory. We currently don't, so there are a lot of tests that get excluded for the chokidar case.

Tracking the task of making them equivalent at https://github.com/google/wireit/issues/550, but figured I'd land the tests first so that we at least don't regress, and have something easy to track for improvements in future.