Fix infinite loop and unnecessary logging in watch mode (#665)
If A depends on B, and B fails but still writes some output that A consumes via its `files` array, then watch mode will currently get stuck in a loop. That's because:

1. Iteration 1 starts
2. B writes an output file, which is also an input file of A
3. Watcher notices a change to an input file so queues another iteration for after this one ends
4. B fails (e.g. there was a type error)
5. Iteration 2 starts
6. B runs again
7. B writes an output file
8. etc...

Step 6 is where the problem is. There's no reason for B to run again. Its input files didn't change, so its fingerprint is the same, and we know that it failed the previous time we tried to run it with the exact same fingerprint. We ran it again because until now we only ever recorded *successful* executions (on disk). There was not a reason to record *failures* before, because when you're in non-watch mode, you would want to retry failures every wireit invocation. So we now do record previous failures, but only in watch mode, and in-memory instead of on disk. If we see that in the previous watch iteration the script had the same fingerprint and failed, then we just immediately fail instead of retrying.

This PR also adds a new `WatchLogger`, which helps reduce unnecessary logging that can come out of the above and similar situations. Previously, every time we started an execution, we'd immediately clear the screen and stream all log messages -- even if all of the log messages were things like "script already fresh". Now we instead buffer events for each watch iteration until we see something that's actually interesting (like running a script), and only then do we clear the screen and start displaying all of the log messages. If in the entire iteration we only saw boring messages, we never show anything.

Thanks @alyahmedaly for the bug report and repro!