Improve security around ACTIONS_RUNTIME_TOKEN (Part 1) (#1146)
Now, instead of exposing the ACTIONS_RUNTIME_TOKEN variable directly, we instead start up a "custodian" HTTP server, which lives for the duration of the workflow, and whose only role is to respond to any HTTP request with the token. We then expose the WIREIT_CACHE_GITHUB_CUSTODIAN_PORT environment variable to the calling-workflow (this provides a positive signal that the custodian is running, and lets us be dynamic with port assignment). Wireit processes can then make a fetch to the custodian service to read the token.