Quick fix for failing windows runs (#1150)
The new GitHub caching setup system (see #1146) is not working on Windows. The Wireit processes are unable to fetch from the local custodian server.

I'm not sure why yet, but in the meantime this release will make it non-fatal when GitHub caching fails to set up.

Also shows more info when a fetch error occurs, and fixes an issue with environment variable inheritance in our integration tests.