Fix GitHub Caching action for Windows (#1157)
Fixes Windows support for GitHub Actions Caching, which has been broken since the security improvement in #1146. (This only affected users who had upgraded to v2 of the action, and the failure was not fatal, it just disabled caching).

The problem was that under Windows, {detached: true} is not sufficient for keeping the server alive after the parent shell exits. The parent shell will exit as soon as the main launching script is done, because the action is invoked through a GitHub Actions uses: clause (as opposed to run: where we would be sharing our parent shell with subsequent steps).

We now use start so that the server is not killed when our parent shell exits: https://learn.microsoft.com/en-us/windows-server/administration/windows-commands/start.

Also listens only on localhost instead of all hosts (shouldn't matter because we should have strong network isolation for free through the GitHub Actions container, but seems wise, and good in case this code runs in some other context for some reason).