Update deps; New eslint config; Fix budgeted fs wrapper bug (#1185)
Update all the dependencies to latest major versions. Only eslint required any changes.

Eslint 9 has a new config file format, so I switched to that. I also adopted the new recommended style for typescript-eslint which has some new rules. The new rules found a few style issues, but also a real issue.

The real issue was that an await was being omitted in the budgeted fs module. We could also update to using, but let's do that in another PR for the whole project.