Support V2 GitHub Caching service (Part 2/2: Wireit) (#1302)
Add support for the new V2 version of the GitHub caching API.

Requires #1301 to land first, which provides a new environment variable we need to read.

Also adds tests against the real production service in CI.

See #1297 and https://github.blog/changelog/2025-03-20-notification-of-upcoming-breaking-changes-in-github-actions/#decommissioned-cache-service-brownouts for more background on the V2 changes.